+++ decidim-proposals/app/helpers/decidim/proposals/control_version_helper.rb
@@ -3,56 +3,43 @@
 module Decidim
   module Proposals
     # Custom helpers, scoped to the proposals engine.
-    #
     module ControlVersionHelper
-      def versions_controller?
-        return true if params[:controller] == "decidim/proposals/versions"
-
-        false
-      end
-
-      def proposal?
-        return true if item.class == Decidim::Proposals::Proposal
-
-        false
+      def item_name
+        item.model_name.singular_route_key.to_sym
       end
 
       def back_to_resource_path_text
-        return unless versions_controller?
-
-        if proposal?
-          t("versions.stats.back_to_proposal", scope: "decidim.proposals")
-        else
-          t("versions.stats.back_to_collaborative_draft", scope: "decidim.proposals")
+        case item_name
+        when :proposal
+          t("back_to_proposal", scope: "decidim.proposals.versions.stats")
+        when :collaborative_draft
+          t("back_to_collaborative_draft", scope: "decidim.proposals.versions.stats")
         end
       end
 
       def back_to_resource_path
-        return unless versions_controller?
-
-        if proposal?
+        case item_name
+        when :proposal
           proposal_path(item)
-        else
+        when :collaborative_draft
           collaborative_draft_path(item)
         end
       end
 
       def resource_version_path(index)
-        return unless versions_controller?
-
-        if proposal?
+        case item_name
+        when :proposal
           proposal_version_path(item, index + 1)
-        else
+        when :collaborative_draft
           collaborative_draft_version_path(item, index + 1)
         end
       end
 
       def resource_all_versions_path
-        return unless versions_controller?
-
-        if proposal?
+        case item_name
+        when :proposal
           proposal_versions_path(item)
-        else
+        when :collaborative_draft
           collaborative_draft_versions_path(item)
         end
       end